Add CASMaxResponseSize.
The default is 65536. Values must be greater than or equal to 0 and less
than or equal to 2147483647.

This replaces the CAS_MAX_RESPONSE_SIZE.

Note that this does not allocate any memory, it simply sets an upper
limit on the number of bytes that a CAS validation response can be.